Move expandable-panels/ to common/
