+++ components/common/expandable-panels/ExpandablePanelSubTitle.module.scss
