+++ components/common/expandable-panels/ExpandableText.module.scss
