+++ components/common/expandable-panels/ExpandableText.tsx
