+++ components/common/index.ts
@@ -1,1 +1,2 @@
 export * from "./card"
+export * from "./expandable-panels";
