+++ components/dotations/articles/dsr-eligibilite/DsrEligibilite.tsx
@@ -1,7 +1,7 @@
 import { Fragment, PureComponent } from "react";
 
 import { StateParameter } from "../../../articles-inputs/parameter";
-import { ExpandableText } from "../../../expandable-panels";
+import { ExpandableText } from "../../../common";
 
 export class DsrEligibilite extends PureComponent {
   render() {
